services/ai-agent-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-anthropic==0.1.0
langchain-community==0.0.10
tavily-python==0.3.0
pymongo==4.6.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.0
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dateutil==2.8.2

